+++ .gitignore
@@ -1,0 +1,2 @@
+# Node modules
+node_modules/
