+++ html/ferramentas.html
@@ -7,9 +7,9 @@
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css">
   <link rel="stylesheet" href="../css/style.css">
   <link rel="stylesheet" href="../css/ferramentas-style.css">
-  <link rel="shortcut icon" href="img/ecomanager.png" type="image/x-icon">
+  <link rel="shortcut icon" href="../img/ecomanager.png" type="image/x-icon">
   <!-- Biblioteca de Gráficos -->
-  <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
+  <!-- <script src="https://cdn.jsdelivr.net/npm/chart.js"></script> -->
   <title>Ferramentas - EcoManager</title>
 </head>
 
@@ -43,7 +43,7 @@
   <header>
     <nav class="navbar" id="navbar">
       <a href="index.html" class="logo">
-        <img src="img/logo-nome.png" alt="EcoManager Logo">
+        <img src="../img/logo-nome.png" alt="EcoManager Logo">
       </a>
 
       <ul class="nav-links" id="navLinks">
@@ -185,12 +185,12 @@
                 </div>
 
                 <div class="result-item">
-                  <span>Consumo de água (Litros):</span>
+                  <span>Total de água consumida:</span>
                   <span class="negrito">0 L</span>
                 </div>
                 
                 <div class="result-item">
-                  <span>Redução no consumo de água (Litros):</span>
+                  <span>Consumo real de água:</span>
                   <span class="negrito">0 L</span>
                 </div>
 
@@ -455,7 +455,7 @@
     </div>
   </footer>
 
-  <script src="../js/script-geral.js"></script>
+  <script type="module" src="../js/mainScript.js"></script>
 </body>
 
 </html>
