+++ html/sobre.html
@@ -5,8 +5,8 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css">
   <!-- Links para os arquivos de estilo -->
-  <link rel="stylesheet" href="css/style.css">
-  <link rel="stylesheet" href="css/sobre-style.css">
+  <link rel="stylesheet" href="../css/style.css">
+  <link rel="stylesheet" href="../css/sobre-style.css">
   <link rel="shortcut icon" href="img/ecomanager.png" type="image/x-icon">
   <title>Sobre Nós - EcoManager</title>
 </head>
