+++ js/scriptCalculadoras/tabs.js
@@ -1,0 +1,14 @@
+export function initCalculatorTabs() {
+  const tabButtons = document.querySelectorAll(".tab-button");
+  const tabContents = document.querySelectorAll(".tab-content");
+
+  tabButtons.forEach(button => {
+    button.addEventListener("click", () => {
+      const targetTab = button.dataset.tab;
+      tabButtons.forEach(btn => btn.classList.remove("active"));
+      tabContents.forEach(content => content.classList.remove("active"));
+      button.classList.add("active");
+      document.getElementById(targetTab)?.classList.add("active");
+    });
+  });
+}
