+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "chart.js": "^4.5.0"
+  }
+}
